backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
openai
python-dotenv
python-multipart
twilio
httpx
pytest
pytest-asyncio
